% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{datasus_file_types}
\alias{datasus_file_types}
\title{List DATASUS file types}
\usage{
datasus_file_types(source = NULL, file_type = NULL)
}
\arguments{
\item{source}{Optional character vector of source codes (e.g. \code{"SIHSUS"}).}

\item{file_type}{Optional character vector of file type codes (e.g. \code{"RD"}).}
}
\value{
A tibble.
}
\description{
Returns the internal catalog of file types. Optionally filtered by source
and/or file type codes.
}
\examples{
datasus_file_types()
datasus_file_types(source = "SIHSUS")
datasus_file_types(source = "CNES", file_type = "ST")
}
