% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{datasus_modalities}
\alias{datasus_modalities}
\title{List DATASUS modalities}
\usage{
datasus_modalities()
}
\value{
A tibble.
}
\description{
Returns the available DATASUS data modalities (data, documentation,
programs, etc.).
}
\examples{
datasus_modalities()
}
