% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_extra_paths}
\alias{.datasus_extra_paths}
\title{Internal FTP paths for territorial and documentation resources}
\usage{
.datasus_extra_paths()
}
\value{
A named list of FTP URLs.
}
\description{
Internal FTP paths for territorial and documentation resources
}
\keyword{internal}
