% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_file_types_tbl}
\alias{.datasus_file_types_tbl}
\title{Internal DATASUS file type catalog}
\usage{
.datasus_file_types_tbl()
}
\value{
A tibble.
}
\description{
Internal DATASUS file type catalog
}
\keyword{internal}
