% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_modalities_tbl}
\alias{.datasus_modalities_tbl}
\title{Internal DATASUS modality catalog}
\usage{
.datasus_modalities_tbl()
}
\value{
A tibble.
}
\description{
Internal DATASUS modality catalog
}
\keyword{internal}
