% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_path_templates_tbl}
\alias{.datasus_path_templates_tbl}
\title{Internal path templates for DATASUS FTP}
\usage{
.datasus_path_templates_tbl()
}
\value{
A tibble.
}
\description{
Internal path templates for DATASUS FTP
}
\keyword{internal}
