% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_programs_tbl}
\alias{.datasus_programs_tbl}
\title{Internal DATASUS program catalog}
\usage{
.datasus_programs_tbl()
}
\value{
A tibble.
}
\description{
Internal DATASUS program catalog
}
\keyword{internal}
