% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasus_catalog.R
\name{.datasus_sources_tbl}
\alias{.datasus_sources_tbl}
\title{Internal DATASUS source catalog}
\usage{
.datasus_sources_tbl()
}
\value{
A tibble.
}
\description{
Internal DATASUS source catalog
}
\keyword{internal}
